requirements.txt
mecab-python3
pytz==2021.1
